CORE-7374 Add "Create Permanent ID" endpoint.
Added a POST /admin/permanent-id-requests/{request-id}/ezid endpoint to
terrain,
which will mint a permanent ID with the EZID API and folder's metadata,
add the new ID(s) to the folder's metadata,
move the folder to the curated directory,
then set the Permanent ID Request's status to "Completed".
If an error is encountered during this process,
then the Permanent ID Request's status will be set to "Failed".
Added EZID API shoulder configs and curated publish directory to
terrain's configs.
Added the EZID API status to the terrain /admin/status response.